docs: update README iron_horus values 3→2 to match code changes

+++ talents/g7/fusion_mixin.py
@@ -5,8 +5,8 @@ class FusionMixin:
     state: Any
     fusion_shield_done: bool
     fusion_weapon_done: bool
-    iron_horus_hp: int
-    iron_horus_max_hp: int
+    iron_horus_hp: float
+    iron_horus_max_hp: float
     eye_of_horus: Any
     ammo: list
     tactical_unlocked: bool
